Refactore transfer object config, removed conatainer, replaced config with proxy
