+++ src/TransferGenerator/Config/ConfigFactoryTrait.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Picamator\TransferObject\TransferGenerator\Config;
 
-use Picamator\TransferObject\TransferGenerator\Config\Container\ConfigContainer;
-use Picamator\TransferObject\TransferGenerator\Config\Container\ConfigInterface;
+use Picamator\TransferObject\TransferGenerator\Config\Config\ConfigInterface;
+use Picamator\TransferObject\TransferGenerator\Config\Config\ConfigProxy;
 
 trait ConfigFactoryTrait
 {
     protected function getConfig(): ConfigInterface
     {
-        return ConfigContainer::getConfig();
+        return new ConfigProxy();
     }
 }
