--- src/TransferGenerator/Config/Container/ConfigContainer.php
@@ -1,34 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Picamator\TransferObject\TransferGenerator\Config\Container;
-
-use Picamator\TransferObject\TransferGenerator\Exception\TransferGeneratorConfigException;
-use Picamator\TransferObject\Generated\ConfigContentTransfer;
-
-class ConfigContainer
-{
-    private static ConfigInterface $config;
-
-    public static function loadConfig(ConfigContentTransfer $configTransfer): void
-    {
-        self::$config = new Config(
-            transferNamespace: $configTransfer->transferNamespace,
-            transferPath: $configTransfer->transferPath,
-            definitionPath: $configTransfer->definitionPath,
-        );
-    }
-
-    /**
-     * @throws \Picamator\TransferObject\TransferGenerator\Exception\TransferGeneratorConfigException
-     */
-    public static function getConfig(): ConfigInterface
-    {
-        if (!isset(self::$config)) {
-            throw new TransferGeneratorConfigException('Configuration not loaded. Run self::loadConfig() first.');
-        }
-
-        return self::$config;
-    }
-}
