+++ src/TransferGenerator/Config/Exception/ConfigNotFoundException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Picamator\TransferObject\TransferGenerator\Config\Exception;
+
+use Exception;
+use Picamator\TransferObject\Exception\TransferExceptionInterface;
+
+class ConfigNotFoundException extends Exception implements TransferExceptionInterface
+{
+}
