+++ src/TransferGenerator/Config/Loader/ConfigLoader.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace Picamator\TransferObject\TransferGenerator\Config\Loader;
 
 use Picamator\TransferObject\Generated\ConfigTransfer;
-use Picamator\TransferObject\TransferGenerator\Config\Container\ConfigContainer;
+use Picamator\TransferObject\TransferGenerator\Config\Config\Config;
+use Picamator\TransferObject\TransferGenerator\Config\Config\ConfigProxy;
 use Picamator\TransferObject\TransferGenerator\Config\Reader\ConfigReaderInterface;
 
 readonly class ConfigLoader implements ConfigLoaderInterface
@@ -19,7 +20,8 @@ readonly class ConfigLoader implements ConfigLoaderInterface
     {
         $configTransfer = $this->reader->getConfig($configPath);
         if ($configTransfer->validator->isValid) {
-            ConfigContainer::loadConfig($configTransfer->content);
+            $config = new Config($configTransfer->content);
+            ConfigProxy::loadConfig($config);
         }
 
         return $configTransfer;
